TreeAutomata: fix unique name generation
The compiler considers calls to `uniqueStart` (and `fresh`) to be
constant expressions, and hence optimizes them away. This results in
calls to these functions always returning the same value: "Start0" (and
"0"). See https://stackoverflow.com/a/19372414 for more information.

The solution described there (using `-O -fno-full-laziness -fno-cse`)
does not seem to work. Hence, we pass a bogus argument to these
functions to force the compiler to not treat them as constants.