Fix `subsetOf` by making grammars deterministic
`subsetOf` missed an edge case where language inclusion is dependend on
more than single production rules. The issue is in the fact that the
inclusion algorithm traverses the productions from the start symbols,
finding equal constructors in both grammars and adding constraints of
the form `S ⊆ S' iff A⊆A'` for two productions `S -> foo(A)` and `S' ->
foo(A')`. For the added test case, this gave rise to the constraints

```
Nt0⊆S iff Nt1⊆G or Nt1⊆G
Nt1⊆G iff Nt2⊆G or Nt2⊆A
Nt2⊆A iff True and False
Nt2⊆G iff (False or False) and (True of False)
```

because the algorithm works "locally", i.e., it only compares one
production rule in one grammar with one production rule in the other.

However, in this case, it is the combination of the production rules
from `Nt1` and `Nt2` that is a subset (equal) to the combination of the
production rules from `G` and `A`. To conclude, what we really ought to
be doing is checking for inclusion not per production, but per a certain
combination of productions.

This is however impossible to do, because what combinations need to be
checked? A quick way out is to require deterministic grammars. In a
deterministic grammar, there are no production rules of the form `X ->
foo(A) | foo(A')` for some nonterminals `X, A` and `A'` and hence, we
can again compare production rules "locally".