determinize: fix space leak by evaluating the GrammarBuilder
Before: `g'` was a variable resulting from evaluating a function.
Everytime we call `evalState` we would also reevaluate `g'`.

Now: `g'` is a thunk, i.e. a piece computation that has not been
executed yet. By refering to `g'` we eventually evaluate it and
subsequent evaluations don't reevaluate it.

In other words, using `evalState` forces the computations to be
completed, and the result is that there is only one new computation and
hence, re-evaluation is prevented.